requirements.txt
streamlit
duckduckgo_search
requests
pandas
plotly
Pillow
alpha_vantage